# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: Nikah Time
  license:
    name: MIT
  version: 0.1.1
servers:
  - url: /api/
paths:
...
  /account/user/update:
    put:
      tags:
        - Account
      summary: Обновление информации о пользователе
      operationId: updateUser
      requestBody:
        description: Информация о пользователе
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "200":
          description: OK
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
  /account/user:
    get:
      tags:
        - Account
      summary: Получение информации о пользователе
      operationId: getUser
      responses:
        "200":
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
...
  /chats/get/message/{messageId}:
    get:
      tags:
        - Chats
      parameters:
        - in: path
          name: messageId
          schema:
            type: integer
          required: true
          description: Айдишник сообщения, данные которого нужно получить
      summary: Получение информации о сообщении пользователя
      operationId: getMessage
      responses:
        "200":
          description: Информация о сообщении между пользователями
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
...
  /search/users:
    post:
      tags:
        - Search
      summary: Поиск пользователей по фильтрам из подбора кандидатур по интересам (из всех сохраненных просмотренных)
      operationId: searchUsers
      requestBody:
        description: Подбор пользователю кандидатур в соответствии с выбранными фильтрами
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filter'
            examples:
              simplifiedFilter:
                value:
                  filterType: simpleFilter
                  minAge: 18
                  maxAge: 30
                  isOnline: true
              complicatedFilter:
                value:
                  filterType: complicatedFilter
                  minAge: 18
                  maxAge: 30
                  country: Россия
                  city: Ростов
                  education: basicGeneral
                  maritalStatus: notMarried
                  haveChildren: true
                  haveBadHabits: false
                  badHabits: ["smoking"]
                  isOnline: true
        required: true
      responses:
        "200":
          description: Кандидатуры пользователей в соответствии с выбранными фильтрами
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favourite'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
  /search/save/seen/users:
    post:
      tags:
        - Search
      summary: Сохранение кандидатур, которые были просмотрены пользователем
      operationId: saveSeenUsers
      requestBody:
        description: Сохранение кандидатур, которые были просмотрены пользователем
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeenUsers'
            example:
              seenUsersId: [1, 2, 3, 4, 5]
        required: true
      responses:
        "200":
          description: Сохранены кандидатуры, которые были просмотрены пользователем
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
  /search/get/selection/users:
    get:
      tags:
        - Search
      summary: Получение кандидатур, которые подобрал сервер в соответствии с интересами пользователя (максимум - 20). Для показа следующих 20 кандидатур, вызываем этот же запрос, здесь будет проверка, были ли уже показаны данные анкеты
      operationId: getSelectionUsers
      responses:
        "200":
          description: Кандидатуры, подобранные пользователю по интересам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favourite'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    Filter:
      required:
        - filterType
        - minAge
        - maxAge
        - isOnline
      type: object
      properties:
        filterType:
          type: string
          description: Тип фильтра
          enum:
            - simplifiedFilter
            - complicatedFilter
        minAge:
          type: integer
          format: int32
        maxAge:
          type: integer
          format: int32
        isOnline:
          type: boolean
        country:
          type: string
        city:
          type: string
        education:
          type: string
          description: "Уровень образования. Может принимать следущие значения: ос\
                  новное общее / basic general, среднее общее / secondary general, средн\
                  ее профессиональное / secondary vocational, высшее (бакалавр, специали\
                  ст) / higher (bachelor, specialist), высшее (магистр) / higher (мaster),\
                  \ учёная степень / academic degree"
          enum:
            - basicGeneral
            - secondaryGeneral
            - secondaryVocational
            - higherBachelorOrSpecialist
            - higherMaster
            - academicDegree
        maritalStatus:
          type: string
          enum:
            - notMarried
            - married
            - divorced
        haveСhildren:
          type: boolean
        badHabits:
          type: array
          items:
            $ref: '#/components/schemas/BadHabit'
        haveBadHabits:
          type: boolean
          description: Обязательный параметр для расширенного фильтра. В случае выбора привычек - передать true
...
    SeenUsers:
      required:
        - seenUsersId
      type: object
      properties:
        seenUsersId:
          type: array
          items:
            type: integer
            format: int64
            description: "Айдишники просмотренных пользователем анкет"
...
    Favourite:
      type: array
      items:
        $ref: '#/components/schemas/FavouriteUser'
...
    User:
      required:
        - userId
        - badHabits
        - birthDate
        - city
        - contactPhoneNumber
        - country
        - education
        - firstName
        - gender
        - haveСhildren
        - interests
        - lastName
        - maritalStatus
        - isOnline
      type: object
      properties:
        userId:
          type: integer
          format: int64
          description: "для прослушивания события отправки сообщения в чате,
                в названии события необходимо будет указывать id авторизованного пользователя"
        firstName:
          type: string
        lastName:
          type: string
        photos:
          type: array
          description: Массив ссылок на фотограйии пользователя. Первая ссылка - основное
            фото пользователя.
          items:
            type: string
            format: uri
        gender:
          type: string
          enum:
            - male
            - female
        birthDate:
          type: string
          format: data-time
          default: 24-09-2021
        country:
          type: string
        city:
          type: string
        contactPhoneNumber:
          type: string
        education:
          type: string
          description: "Уровень образования. Может принимать следущие значения: ос\
            новное общее / basic general, среднее общее / secondary general, средн\
            ее профессиональное / secondary vocational, высшее (бакалавр, специали\
            ст) / higher (bachelor, specialist), высшее (магистр) / higher (мaster),\
            \ учёная степень / academic degree"
          enum:
            - basicGeneral
            - secondaryGeneral
            - secondaryVocational
            - higherBachelorOrSpecialist
            - higherMaster
            - academicDegree
        placeOfStudy:
          type: string
        placeOfWork:
          type: string
        workPosition:
          type: string
        maritalStatus:
          type: string
          enum:
            - notMarried
            - married
            - divorced
        haveСhildren:
          type: boolean
        badHabits:
          type: array
          items:
            $ref: '#/components/schemas/BadHabit'
        interests:
          type: array
          items:
            type: string
        about:
          type: string
        isOnline:
          type: boolean
    FavouriteUser:
      required:
        - userId
        - badHabits
        - birthDate
        - city
        - country
        - education
        - firstName
        - haveСhildren
        - interests
        - lastName
        - maritalStatus
        - isOnline
      type: object
      properties:
        userId:
          type: integer
          format: int64
          description: "Нужен для добавления данного пользователя в понравившиеся и для начала чата с данным пользователем"
        firstName:
          type: string
        lastName:
          type: string
        photos:
          type: array
          description: Массив ссылок на фотограйии пользователя. Первая ссылка - основное
            фото пользователя.
          items:
            type: string
            format: uri
        birthDate:
          type: string
          format: data-time
          default: 24-09-2021
        city:
          type: string
        education:
          type: string
          description: "Уровень образования. Может принимать следущие значения: ос\
              новное общее / basic general, среднее общее / secondary general, средн\
              ее профессиональное / secondary vocational, высшее (бакалавр, специали\
              ст) / higher (bachelor, specialist), высшее (магистр) / higher (мaster),\
              \ учёная степень / academic degree"
          enum:
            - basicGeneral
            - secondaryGeneral
            - secondaryVocational
            - higherBachelorOrSpecialist
            - higherMaster
            - academicDegree
        placeOfStudy:
          type: string
        placeOfWork:
          type: string
        workPosition:
          type: string
        maritalStatus:
          type: string
          enum:
            - notMarried
            - married
            - divorced
        haveСhildren:
          type: boolean
        badHabits:
          type: array
          items:
            $ref: '#/components/schemas/BadHabit'
        interests:
          type: array
          items:
            type: string
        about:
          type: string
        isOnline:
          type: boolean
...
    BadHabit:
      type: string
      enum:
        - smoking
        - alcohol
        - gambling
        - other
    Error:
      required:
        - code
        - detail
        - title
      type: object
      properties:
        code:
          type: integer
          format: int32
        title:
          type: string
        detail:
          type: string
...
    Message:
      required:
        - message
        - messageTime
        - isAuthUserMessage
        - messageId
        - isMessageSeen
        - messageType
      type: object
      properties:
        message:
          type: string
        messageTime:
          type: string
          format: data-time
        isAuthUserMessage:
          type: boolean
        messageId:
          type: integer
          format: int64
        isMessageSeen:
          type: boolean
        messageType:
          type: string
          description: Тип сообщения, например, картинка, текст и т.д
          enum:
            - text
            - image
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT